high level refactoring scopedvariableservice
